Set 'executable' for JavaExec tasks
For now I don't know if it's possible to find the JavaCompile task which JavaExec depends on.
I have tried with task.dependsOn and task.taskDependencies but they are empty.

The plugin just tries to find all Java version used in JavaCompile tasks and do nothing if there
is more that one version found 🤷‍♂️.